Removed the SECOND var
There's no need to make a variable called SECOND if you're only going to use it once. Also, it's pretty negligible, but the name of the variable is longer than the amount of characters in the value so performance wise its actually worse. As I said, it's pretty much negligible, but if you keep doing it, it adds up. 

It's fine if you don't accept the merge, just something to think about.